backend/requirements.txt
fastapi
uvicorn
transformers
torch
python-dotenv
fastapi[all]
sqlalchemy
passlib[bcrypt]
python-jose